refactor(forgetest): Split integration harness
The forgetest integration harness keeps credentials, HTTP recording, suite
orchestration, scenario bodies, local Git setup, and random data generation in
one large file. Split those responsibilities into package-local files so each
maintenance path has a smaller surface to inspect.

This is a file organization change inside internal/forge/forgetest. Exported
APIs, fixture names, subtest names, provider callbacks, and scenario behavior
remain unchanged.
